openrouter_module/pyproject.toml
[project]
name = "openrouter-playground"
version = "0.1.0"
description = "Simple app for testing and exploring OpenRouter API"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = []

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"